Import of getFurthestNonVerseInline is broken- move function into files that use it
